backend/requirements.txt
Django>=5.0
psycopg2-binary>=2.9
python-decouple>=3.8
django-ninja>=0.18.0
django-cors-headers>=4.0
celery>=5.3
redis>=5.0 
dj-database-url==3.0.0
pydantic
pydantic[email]
pyjwt>=2.0
pydantic-ai>=0.0.14
anthropic>=0.21.0
requests>=2.31.0
django-extensions==4.1
ipdb==0.13.13
gunicorn>=21.0
whitenoise>=6.6